Auth updates

+++ src/NetApi/Models/Dtos/Paginated.cs
@@ -3,7 +3,7 @@ namespace NetApi.Models.Dtos;
 public class Paginated<T>
 {
     public IEnumerable<T> Items { get; set; } = [];
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
     public int PageSize { get; set; }
-    public int StartIndex { get; set; } 
+    public long StartIndex { get; set; }
 }
